Extracted utils.create_new_sg_entity to standardise logic in full and update sync
